rc2nix: Skip settings that are dependent on state
This change ensures that settings that are related to state are not
exported as Nix options.  This ensures that settings can be applied on
systems with different states.

For example, settings that are attached to specific Plasma widgets,
color settings that come from themes, or settings relating to window
sizes or positions.