rc2nix: Properly handle shortcuts with no defined keys (fixes #6)
If a shortcut has no keys defined then splitting on the key definition
string will return an empty array, which causes `first` to return
`nil` and all hell breaks lose.  Empty key definitions are suppose to
be set to "none", but if that isn't the case we should handle it.