fix(startup): use heredocs instead of quoted strings for desktop scripts
If a desktop script contains repeated ' (single apostrophes), the escape
function would produce a result that's invalid within double quotes.

To solve this, we use `read` to read the contents of the script from a
heredoc, and put it into a variable which can be quoted, which then
eliminates the issue.