modules(panels/widgets): add digital clock widget
Adds settings for the digital clock widget.

Settings are directly sourced from KDE's documentation
(https://develop.kde.org/docs/plasma/scripting/keys/), so it should be
fairly comprehensive. The settings are arranged in a way that fits the
configuration GUI better, as well as providing some structure to the
smattering of options in the raw config file.