modules(widgets): refactor to use the module system more
Someone recently showed me that there's an `apply` setting for mkOption
that automagically does the type conversions that we've done manually so
far. This is great news as it cut down redundant code massively, and
we're able to extract common patterns like enums into the widgets lib.

I also reworked system tray's item configs to let it use the option
system to fill in the defaults *as specified* by the options. This
allows us to remove a lot of duplicate handling code that are designed
to fill in sensible defaults even when it's not been processed by the
option system.

Converters are frankly *much* easier to understand now, which is always a
good thing in my book.