Fix font configuration for Plasma 5
Before the fix, setting the fonts results in a parsing error and the
effective font reverting to a default instead.
When running a program, the error is logged by QFont::fromString() in Qt5.
Refer to:
https://code.qt.io/cgit/qt/qtbase.git/tree/src/gui/text/qfont.cpp?h=5.15#n2156
Notice that it will only accept 10 items in the font string
(excluding optional style name). Compare that to the Qt6 (for Plasma 6):
https://code.qt.io/cgit/qt/qtbase.git/tree/src/gui/text/qfont.cpp?h=6.7.1#n2182
we see that the unsupported fields we need to remove are the last 6 of them.

Additionally, the font weight enum changed in Qt6, so we have to fix it
be the Qt5 scale.
Qt 5: https://doc.qt.io/qt-5/qfont.html#Weight-enum
Qt 6: https://doc.qt.io/qt-6/qfont.html#Weight-enum