feat(cursors): Cursor launch feedback (#471)
* Added cursorFeedback entry to cursorType

This is to assist in implementing declarable launch feedback settings in
kde.

* Implemented cursor launch feedback output to klaunchrc

Implementation does not match exact behaviours of kde settings app. For
example, when selecting the "Bouncing" option for cursor feedback kde
will have no settings written to klaunchrc.

However this implementation is very explicit and outputs all cursor settings
options. Although it differs from kde's implementation, this more explicit
version still works.

* Added taskManagerFeedback and animationTime entries to cursorType

This is to assist in making the feedback displayed on the icon in the
task manager a declerative setting.

* Added configuring of taskManagerFeedback and animationTime

* Refactored klaunchrc configuration

* Formatted using nixfmt

* Changed cursorFeedback to use enum instead of string

Now the allowed values are enforced when configuring cursor feedback.

* Removed unnecessary quotations on booleans

* Grouped busy cursor settings